requirements.txt
albucore==0.0.11
albumentations==1.4.10
annotated-types==0.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
idna==3.4
imageio==2.34.1
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.2.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.28.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.6.18
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchinfo==1.8.0
torchvision==0.16.2+cu118
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.9.0
urllib3==1.26.13
